requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
black==24.2.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-ninja==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
ecdsa==0.18.0
filelock==3.13.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
pyasn1==0.5.1
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-jose==3.3.0
pytz==2024.1
rsa==4.9
setuptools==69.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
virtualenv==20.25.0
wheel==0.42.0
